+++ src/main/java/br/com/jvsmed/api/entities/Medico.java
@@ -43,16 +43,11 @@ public class Medico {
     @Column(name = "crm", length = 6, nullable = false)
     private String crm;
 
-    private Boolean ativo;
-
     public Medico(DadosCadastroMedico dados) {
-        this.ativo = true;
         this.nome = dados.nome();
         this.email = dados.email();
         this.telefone = dados.telefone();
         this.crm = dados.crm();
-        //this.especialidade = dados.especialidade();
-        //this.endereco = new Endereco(dados.endereco());
     }
 
     public void atualizarInformacoes(DadosAtualizacaoMedico dados) {
@@ -62,10 +57,8 @@ public class Medico {
         if (dados.telefone() != null) {
             this.telefone = dados.telefone();
         }
-
-    }
-
-    public void inativar() {
-        this.ativo = false;
+        if (dados.email() != null) {
+            this.email = dados.email();
+        }
     }
 }
