+++ src/main/java/br/com/jvsmed/api/enums/ECategoria.java
