+++ src/main/java/br/com/jvsmed/api/enums/ESolicitacao.java
