--- src/main/java/br/com/jvsmed/api/medico/MedicoRepository.java
@@ -1,11 +1,0 @@
-package br.com.jvsmed.api.medico;
-
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Range;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MedicoRepository extends JpaRepository<Medico, Long> {
-
-    Page<Medico> findAllByAtivoTrue(Pageable paginacao);
-}
