--- src/main/java/br/com/jvsmed/api/paciente/DadosAtualizacaoPaciente.java
@@ -1,33 +1,0 @@
-package br.com.jvsmed.api.paciente;
-
-import br.com.jvsmed.api.cnh.Solicitacao;
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotNull;
-
-public record DadosAtualizacaoPaciente(
-        @NotNull
-        Long id,
-        String guia,
-        String registro,
-        String categoria,
-        String data_cadastro,
-        String data_habilitacao,
-        String nome,
-        String data_nascimento,
-        String sexo,
-        String cpf,
-        String nacionalidade,
-        String nome_mae,
-        String nome_pai,
-        String telefone,
-        String logradouro,
-        String bairro,
-        String cep,
-        String numero,
-        String complemento,
-        String cidade,
-        String uf
-
-) {
-
-}
