--- src/main/java/br/com/jvsmed/api/paciente/PacienteRepository.java
@@ -1,14 +1,0 @@
-package br.com.jvsmed.api.paciente;
-
-import br.com.jvsmed.api.medico.Medico;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface PacienteRepository extends JpaRepository<Paciente, Long> {
-    Page<Paciente> findAllByAtivoTrue(Pageable paginacao);
-
-    Paciente findByCpf(String cpf);
-}
