--- src/main/java/br/com/jvsmed/api/paciente/Sexo.java
@@ -1,9 +1,0 @@
-package br.com.jvsmed.api.paciente;
-
-public enum Sexo {
-    MASCULINO("Masculino"),
-    FEMININO("Feminino");
-
-    Sexo(String sexo) {
-    }
-}
