+++ src/main/java/br/com/jvsmed/api/registro/medico/DadosAtualizacaoMedico.java
@@ -7,6 +7,7 @@ public record DadosAtualizacaoMedico(
         @NotNull
         Long id,
         String nome,
-        String telefone,
-        DadosEndereco endereco) {
+        String email,
+        String telefone
+        ) {
 }
